+++ RadioControl/app/src/main/java/com/nikhilparanjape/radiocontrol/rootUtils/RootAccess.java
@@ -24,8 +24,7 @@ public class RootAccess{
         }
         catch (IOException e){
             e.printStackTrace();
-            FirebaseCrash.logcat(Log.ERROR, "RadioControl", "Root Error");
-            FirebaseCrash.report(e);
+
         }
     }
 }